chore: Change development to a single flox env
This commit changes to use a single flox env instead of several in each
subproject of the monorepo. The idea is that we have conventions for go and
node projects and they can be specified via a single enviornment that works in
all subprojects.

This commit adds a toplevel flox environment that can be activated using `flox
activate` once you have flox installed from https://flox.dev.

This also cleans up gitignore to move most of the logic into the top level
gitignore with less in the others.