feat(sn): initial implementation of web link
Previously there was no link to the zoom meeting that the zoom-notifier was
sending messages about. This changeset uses rich text formatting via the slack
block builder API to inline a link to the zoom meeting if the proper settings
are provided. These settings are defined in environment variables, but also
require setup on the Zoom application side.

A future commit will have full documentation. This commit is the simple MVP
with a lot of debugging enabled still. This is basically the happy path.